Create section for installing from private package repos (#597)
The approach is very similar to installing from a GH repo, but in this case we are installing prebuilt packages that are hosted on a private package repository. This was already asked in ZD 8829 and so I think it would be good to show how this is done. I think we will see more and more people switching towards prepackaged packages because it's a better way of distributing packages - however until Github do it, it might not become that popular (github/roadmap#94)