clarity around aws secrets manager implementation (#737)
* aws secrets manager has incorrect characters

backtick on line 385 causes example to not work, also adding quotes around values to substitute since they are required in Dockerfile & .env

* clarity around aws secrets manager implementation

customer ran into several days of troubleshooting due to lack of clarity around leading `/` in variable names - I am udpating the `AIRFLOW__SECRETS__BACKEND_KWARGS` to match this, as the prefix setting doesn't discern between using leading `/` or not, but setting the actual secret names will not work if a leading `/` is used.  

Also tried to clarify a few points but unsure if wording or formatting is most ideal. Definitely open to other suggestions, and if some of these are approved we may want to propagate the same changes to the other providers

* adding option to deploy directly from `.env`

- now that `astrocloud deployment variable update` command is available via the CLI , this option should be preferred over using Dockerfile for testing ENV Vars as it requires additional steps to ensure secrets are removed and not deployed
- added some clarity and info/warnings for a few sections


Please let me know your thoughts on these changes and feel free to make any adjustments as needed. If these generally look good, we should apply similar changes to the other secrets backend options

* Update astro/secrets-backend.md

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>

* Update astro/secrets-backend.md

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>

* Update astro/secrets-backend.md

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>

* Update astro/secrets-backend.md

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update astro/secrets-backend.md

Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>

* Update secrets-backend.md

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>
Co-authored-by: Paola Peraza Calderon <paola@astronomer.io>