Remove AWS perms, clarify position on audit (#1541)
* Remove permissions from this document 

Feedback has been that having each permission listed in this document is too AWS-centric (i.e. we are basically regurgitating what each API family does on AWS).  Instead, we will point users to the policy in the public S3 bucket and they can examine this on their own. 

Also clarified that Astro cannot modify the CloudTrail that they set up, so there's no chance we can mess with their audit.

* Apply suggestions from code review

Co-authored-by: Rob Gray <104205257+kwcanuck@users.noreply.github.com>