improve Astronomer Software airgapped release service instructions
Previous instructions wer unclear as to what nginx image to use
and expected users to provide their own nginx image. This commit
updates the instructions to use the same image that we already
ship for use with astronomer-nginx-es and adjusts the containerPort
and targetPort to use 8080 which is what our image expects.

This is a drop-in replacement for any existing users, the service
port remains unchanged.